+++ CHANGES.md
@@ -24,6 +24,7 @@ to use the Prism Scanner without Rails support.
 - Loads environment variables via `dotenv` if available. [#395](https://github.com/glebm/i18n-tasks/issues/395)
 - Adds CLI command `check_prism` to try the new parser out and see the differences in key detection.
 - The Prism-based scanner supports candidate_keys for Rails translations, allowing relative translations in controllers to match either the key scoped to controller and action or only to the controller.
+- Translation services now catch errors and save partial results [#642](https://github.com/glebm/i18n-tasks/issues/642)
 
 ## v1.0.15